Evaluating selection accuracy based on the MCAA Framework

PartInv, PartInvMulti_we evaluate partial measurement invariance using an extension of Millsap & Kwok's (2004) approach

Usage

PartInvMulti_we(
  propsel,
  cut_z = NULL,
  weights_item = NULL,
  weights_latent = NULL,
  kappa_r = NULL,
  kappa_f = kappa_r,
  alpha_r,
  alpha_f = alpha_r,
  phi_r = NULL,
  phi_f = phi_r,
  psi_r,
  psi_f = psi_r,
  lambda_r,
  lambda_f = lambda_r,
  tau_r = NULL,
  tau_f = tau_r,
  nu_r,
  nu_f = nu_r,
  Theta_r,
  Theta_f = Theta_r,
  pmix_ref = 0.5,
  plot_contour = FALSE,
  show_mi_result = FALSE,
  labels = c("Reference", "Focal"),
  ...
)

PartInv(
  propsel,
  cut_z = NULL,
  weights_item = NULL,
  weights_latent = NULL,
  kappa_r = NULL,
  kappa_f = kappa_r,
  alpha_r,
  alpha_f = alpha_r,
  phi_r = NULL,
  phi_f = phi_r,
  psi_r,
  psi_f = psi_r,
  lambda_r,
  lambda_f = lambda_r,
  tau_r = NULL,
  tau_f = tau_r,
  nu_r,
  nu_f = nu_r,
  Theta_r,
  Theta_f = Theta_r,
  pmix_ref = 0.5,
  plot_contour = FALSE,
  show_mi_result = FALSE,
  labels = c("Reference", "Focal"),
  ...
)

Arguments

propsel

Proportion of selection. If missing, computed using cut_z.

cut_z

Pre-specified cutoff score on the observed composite. This argument is ignored when propsel has input.

weights_item

A vector of item weights.

weights_latent

A vector of latent factor weights.

alpha_r

A vector of latent factor means for the reference group.

alpha_f

(optional) A vector of latent factor means for the focal group; if no input, set equal to alpha_r.

phi_r, phi_f, tau_r, tau_f, kappa_r, kappa_f

Deprecated; included only for backward compatibility.

psi_r

A matrix of latent factor variance-covariances for the reference group.

psi_f

(optional) A matrix of latent factor variance-covariances for the focal group; if no input, set equal to psi_r.

lambda_r

A matrix of factor loadings for the reference group.

lambda_f

(optional) a matrix of factor loadings for the focal group; if no input, set equal to lambda_r.

nu_r

A matrix of measurement intercepts for the reference group.

nu_f

(optional) A matrix of measurement intercepts for the focal group; if no input, set equal to nu_r.

Theta_r

A matrix of the unique factor variances and covariances for the reference group.

Theta_f

(optional) A matrix of the unique factor variances and covariances for the focal group; if no input, set equal to Theta_r.

pmix_ref

Proportion of the reference group; default to 0.5 (i.e., two populations have equal size).

plot_contour

Logical; whether the contour of the two populations should be plotted; default to TRUE.

show_mi_result

If TRUE, perform selection accuracy analysis for both the input parameters and the implied parameters based on a strict invariance model, with common parameter values as weighted averages of the input values using pmix_ref.

labels

A character vector with two elements to label the reference and the focal group on the graph.

...

Other arguments passed to the contour function.

Value

The output will be a list of four elements and a plot if plot_contour == TRUE:

propsel

Echo the same argument as input.

cutpt_xi

Cut point on the latent scale (xi).

cutpt_z

Cut point on the observed scale (Z).

summary

A 8 x 3 table, with columns representing the reference, the focal, and the expected results if the latent distribution of focal group matches the reference group. The rows represent probabilities of true positive (A), false positive (B), true negative (C), false negative (D); proportion selected, success ratio, sensitivity, and specificity.

Examples

# Single dimension
PartInv(propsel = .30,
        weights_item = c(1, 1, 1, 1),
        weights_latent = 1,
        alpha_r = 0,
        alpha_f = 0,
        psi_r = 1,
        lambda_r = c(1, 1, 1, 1),
        nu_r = c(1, 1, 1, 2),
        nu_f = c(1, 1, 1, 1),
        Theta_r = diag(1, 4),
        labels = c("Female", "Male"),
        show_mi_result = FALSE)
#> Proportion selected:  0.3 
#> Cutpoint on the latent scale (xi):  0.524 
#> Cutpoint on the observed scale (Z):  6.86 
#> AI ratio:  0.771 
#> 
#> Classification Accuracy Indices:
#>                     Female  Male E_R(Male)
#> True Positive        0.252 0.217     0.217
#> False Positive       0.087 0.045     0.045
#> True Negative        0.613 0.655     0.655
#> False Negative       0.048 0.083     0.083
#> Proportion Selected  0.339 0.261     0.261
#> Success Ratio        0.743 0.829     0.829
#> Sensitivity          0.839 0.722     0.722
#> Specificity          0.876 0.936     0.936
 res <- PartInv(propsel = .50,
        alpha_r = 0.5,
        alpha_f = 0,
        psi_r = 1,
        lambda_r = c(1, 1, 1, 1),
        nu_r = c(1, 2, 1, 2),
        nu_f = c(1, 1, 1, 1),
        Theta_r = diag(1, 4),
        labels = c("Female", "Male"),
        show_mi_result = TRUE)
# multiple dimensions
lambda_matrix <- matrix(0, nrow = 5, ncol = 2)
lambda_matrix[1:2, 1] <- c(.322, .655)
lambda_matrix[3:5, 2] <- c(.398, .745, .543)
PartInv(propsel = .05,
        weights_latent = c(0.5, 0.5),
        alpha_r = c(0, 0),
        alpha_f = c(-0.3, 0.1),
        psi_r = matrix(c(1, 0.5, 0.5, 1), nrow = 2),
        lambda_r = lambda_matrix,
        nu_r = c(.225, .025, .010, .240, .125),
        nu_f = c(.225, -.05, .240, -.025, .125),
        Theta_r = diag(1, 5),
        Theta_f = c(1, .95, .80, .75, 1))
#> Proportion selected:  0.05 
#> Cutpoint on the latent scale (xi):  1.377 
#> Cutpoint on the observed scale (Z):  5.765 
#> AI ratio:  0.859 
#> 
#> Classification Accuracy Indices:
#>                     Reference Focal E_R(Focal)
#> True Positive           0.023 0.018      0.022
#> False Positive          0.033 0.026      0.026
#> True Negative           0.911 0.930      0.918
#> False Negative          0.033 0.026      0.034
#> Proportion Selected     0.056 0.044      0.048
#> Success Ratio           0.416 0.409      0.458
#> Sensitivity             0.415 0.410      0.393
#> Specificity             0.965 0.973      0.972
PartInvMulti_we(propsel = .10,
                weights_item = c(1/3, 1/3, 1/3, 1/3),
                weights_latent = 1,
                alpha_r = 0.5,
                alpha_f = 0,
                psi_r = 1,
                lambda_r = c(.3, .5, .9, .7),
                nu_r = c(.225, .025, .010, .240),
                nu_f = c(.225, -.05, .240, -.025),
                Theta_r = diag(.96, 4),
                labels = c("female", "male"),
                show_mi_result = TRUE)
#> Proportion selected:  0.1 
#> Cutpoint on the latent scale (xi):  1.572 
#> Cutpoint on the observed scale (Z):  1.701 
#> AI ratio:  0.944 
#> 
#> Classification Accuracy Indices:
#>                     female  male E_R(male)
#> True Positive        0.080 0.030     0.077
#> False Positive       0.056 0.035     0.051
#> True Negative        0.802 0.907     0.807
#> False Negative       0.062 0.028     0.065
#> Proportion Selected  0.136 0.064     0.128
#> Success Ratio        0.587 0.461     0.600
#> Sensitivity          0.562 0.509     0.543
#> Specificity          0.935 0.963     0.940