Compute the mean, standard deviation, and covariance of latent and observed variables.

mn_sd_cov is a helper function that computes the mean, standard deviation, and covariance of latent and observed variables for each group.

Usage

mn_sd_cov(weights_item, weights_latent, alpha, psi, lambda, nu, theta)

Arguments

weights_item

A vector of item weights.

weights_latent

A vector of latent factor weights.

alpha

A list of length g containing 1 x d latent factor mean vectors where g is the number of groups and d is the number of latent dimensions. The first element is assumed to belong to the reference group.

psi

A list of length g containing d x d latent factor variance-covariance matrices where g is the number of groups and d is the number of latent dimensions. The first element is assumed to belong to the reference group.

lambda

A list of length g containing n x d factor loading matrices where g is the number of groups, d is the number of latent dimensions, and n is the number of items in the scale. The first element is assumed to belong to the reference group.

nu

A list of length g containing 1 x n measurement intercept vectors where g is the number of groups and n is the number of items in the scale. The first element is assumed to belong to the reference group.

theta

A list of length g containing 1 x n vectors or n x n matrices of unique factor variances and covariances, where g is the number of groups and n is the number of items in the scale. The first element is assumed to belong to the reference group.

Value

The output will be a list of 5 elements:

mn_z

Mean of the observed variable.

sd_z

Standard deviation of the observed variable.

mn_xi

Mean of the latent variable.

sd_xi

Standard deviation of the latent variable.

cov_z_xi

Covariance of the latent and observed variables.